requirements.txt
django
python-decouple
pillow
djangorestframework
drf-yasg